+++ lib/src/pages/bloc_material.dart
@@ -9,8 +9,8 @@ import 'material.dart';
 /// [P] is the base class for all app's page path.
 /// [R] is the result returned when the page pops.
 /// [B] is the block for the state.
-class CBlocMaterialPage<P extends PagePath, R,
-    B extends CPageBloc<P, R>> extends CAbstractMaterialPage<P, R> {
+class CBlocMaterialPage<P extends PagePath, R, B extends CPageBloc<P, R>>
+    extends CAbstractMaterialPage<P, R> {
   @override
   final B bloc;